pyproject.toml
[project]
name = "fastfood-api"
version = "0.1.0"
description = "Uma api de fastfood para estudar arquitetura hexagonal modular"
readme = "README.md"
authors = [
    { name="Filipe Braga", email="developer@prokicon.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dependency-injector>=4.46.0",
    "fastapi[all]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pyright>=1.1.399",
    "ruff>=0.11.5",
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
lint = "ruff check src"
fix = "ruff check src --fix && black src"
typecheck = "pyright src"
run_dev = "fastapi dev src/main.py"

